Feature/review

+++ src/main/java/com/nhnacademy/bookpubshop/review/repository/impl/ReviewRepositoryImpl.java
@@ -255,7 +255,7 @@ public class ReviewRepositoryImpl extends QuerydslRepositorySupport
                 .select(Projections.constructor(
                         GetProductReviewInfoResponseDto.class,
                         review.count().coalesce(0L).as("reviewCount"),
-                        review.reviewStar.avg().longValue().coalesce(0L).as("productStar")
+                        review.reviewStar.avg().intValue().coalesce(0).as("productStar")
                 ))
                 .where(review.product.productNo.eq(productNo)
                         .and(review.reviewDeleted.isFalse()))
